* libmutil: Fixes for compile time warnings.
git-svn-id: svn://svn.minisip.org/minisip/trunk@852 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/source/mtime.cxx
@@ -41,6 +41,6 @@ uint64_t mtime(){
 
 	gettimeofday( &tv, NULL );
 
-	return ((uint64_t)tv.tv_sec) * 1000LL + ((uint64_t)tv.tv_usec) / 1000LL;
+	return ((uint64_t)tv.tv_sec) * (uint64_t)1000 + ((uint64_t)tv.tv_usec) / (uint64_t)1000;
 #endif
 }
